+++ src/main/java/com/task/decisionengine/domain/UserProfileRegistry.java
@@ -1,5 +1,5 @@
 package com.task.decisionengine.domain;
 
 interface UserProfileRegistry {
-    double findUserCreditModifierByPersonalCode(String personalCode);
+    int findUserCreditModifierByPersonalCode(String personalCode);
 }
